server router fix
