+++ lib/moviesLib.js
@@ -7,17 +7,17 @@ export const getUpcomingMovies = async () => {
 }
 
 export const searchMovies = async (query) => {
-    const {data} = await axios.get(`/movie/search/${query}`)
+    const {data} = await axios.get(`/search/movie/${query}`)
     return data
 }
 
 export const searchTVshows = async (query) => {
-    const {data} = await axios.get(`/movie/tv/${query}`)
+    const {data} = await axios.get(`/search/tv/${query}`)
     return data
 }
 
 export const searchCast = async (query) => {
-    const {data} = await axios.get(`/movie/cast/${query}`)
+    const {data} = await axios.get(`/search/cast/${query}`)
     return data
 }
 
@@ -25,12 +25,12 @@ export const searchCast = async (query) => {
 
 
 export const getMovieDetails = async (id) => {
-    const {data} = await axios.get(`/movie/movieID/${id}`)
+    const {data} = await axios.get(`/movie/${id}`)
     return data
 }
 
 export const getTVshowDetails = async (id) => {
-    const {data} = await axios.get(`/movie/tvID/${id}`)
+    const {data} = await axios.get(`/tv/${id}`)
     return data
 }
